[refactor] Fold linkToUser into the courseRegistrations router
